+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,6 +1,7 @@
 Version X.X.X
 =============
 
+ * 2014-07-10 Added support for forbidden error to Ninja default results. (ra)
  * 2014-06-06 Fixed testcase that was flaky in GMT-5 timezones. (ra) 
  * 2014-06-05 Fixed dependency problems with scope test and ninja-test-utilities (ra)